+++ src/main/java/com/project/CineMe_BE/mapper/response/MovieResponseMapper.java
@@ -31,11 +31,11 @@ public interface MovieResponseMapper extends BaseResponseMapper<MovieResponse, M
 
 
     @Named("mapRating")
-    default double mapRating(Set<ReviewEntity> listReview) {
-        return listReview.stream()
+    default int mapRating(Set<ReviewEntity> listReview) {
+        return (int) listReview.stream()
                 .mapToDouble(rating -> rating.getRating())
                 .average()
-                .orElse(1.0) / 10 * 100;
+                .orElse(1) / 10 * 100;
     }
